fix(logs): always shwo logs tab for resources
The logs tab was sometimes not shown when a custom logs query was used.
This happened for nodes and namespaces and is now fixed by always
showing the logs tab.
